% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{follower_matrix}
\alias{follower_matrix}
\title{Construct a Twitter follower matrix}
\usage{
follower_matrix(users, method = "friends")
}
\arguments{
\item{users}{A character vector of Twitter handles (with no '@'
symbol) or user IDs}

\item{method}{Construct the matrix by gathering either "friends" or
"followers" data for each account. "friends" will be faster for
accounts with more followers than friends, and vice versa.}
}
\value{
A square matrix representing the directed graph of twitter
follower relationships. Each entry in the matrix is 1 if the row
account is following the column account, otherwise 0.
}
\description{
Collect data from the Twitter API and create a follower matrix
}
\examples{
\dontrun{
setup_twitter_oauth(consumer_key = "[your info]",
    consumer_secret = "[your info]",
    access_token = "[your info]",
    access_secret = "[your info]")
users = c("c4ssdotorg", "RoderickTLong", "AynRandInst", "AynRandOrg")
m = follower_matrix(users)

# plot with igraph
library(igraph)
g = graph_from_adjacency_matrix(m, "directed")
plot(g)
}
}

